requirements.txt
discord.py
os
random
re
nltk
